MetadataManager: make it public
Even if it's not ideal, in Gtef the API doesn't need to be perfect.

Another solution would be to call the init() and shutdown() functions
internally, but it would suppose that GApplication is used.
init() would be called with a path to
~/.cache/gtksourceview/<app-id>/metadata.xml with <app-id> the
GApplication ID, but it doesn't always exist. Or take the program
name with g_get_prgname(), or whatever.
The MetadataManager would not support if several processes try to
read/write the same *.xml file, that's why having a separate *.xml file
for each process is needed. With GApplication process uniqueness, we can
take the GApplication ID.

But it's simpler to just expose init() and shutdown(), so that an
application can call those functions itself. API-wise, it's not perfect,
but it's simpler.