tab: add test
It's the first interactive test.

When there will be all the different file loading and saving error info
bars, it will be possible to show each info bar from test-tab. Because
reproducing each error in a real application can be difficult.

Anyway, even if GtefTab doesn't contain a lot of code, it's better to be
able to test each change early on.