Factor out GtefFileMetadata
Move metadata stuff to its own class, to have small classes that do one
thing but do them well. It's in preparation to merge GtkSourceFile into
GtefFile (to basically fork the GtkSourceFile* API, which will allow
complete freedom to refactor the code, temporarily removing features to
start with a simple code, and then re-adding more features afterwards,
and do the necessary refactorings to have a higher-level API, without
the need to adapt gedit).