GutterRendererFolds: remove LINE_WIDTH #define
I think that I remember that I wanted to make the line width
configurable, but the functions that do the drawings do not take
LINE_WIDTH into account, the code was in an unfinished state.

But I don't think being able to configure the line width is something
useful. So, remove the #define (changing its value anyway didn't work).