FoldRegion: some fixes, minor coding style changes and add FIXMEs
- dispose is better than finalize, to release other GObjects (to break
ref cycles, if any). When creating a GtkTextMark (even private), it is
possible for some other code to get it via a signal, and another class
could ref it (it would probably be wrong to do so, but some apps have a
plugin system, so a third-party plugin could cause some problems
elsewhere in the app, which we want to avoid as much as possible). When
creating a truly private GObject, then it's ok to unref it in finalize,
if it is not exposed in the public API of the class.

- some docs improvements

- fix return value in a g_return_val_if_fail(): FALSE -> NULL. Both have
the same value (0), but it is better to use the right "zero", for better
code readability.

- a g_return_if_fail() was missing.

- add early returns to remove one indentation level.