FileSaver: do not use an heuristic to know if it's a "save as" operation
With a GtkFileChooserDialog, it is possible to save as to the same
location as previously set. There is then a confirmation dialog, so what
the user wants is to replace the file. So the heuristic was not 100%
correct, the prev_location can be the same. It can be useful to save
with another encoding or another line ending type.

A "save as" operation is always performed with
gtef_file_saver_new_with_target(). So it's the best place to configure
the FileSaver correctly for a "save as" operation.