File: check_file_on_disk(): set deleted to FALSE if the file exists
Maybe the file has been deleted, and then a new file with the same name
exists again, so it'll be marked as externally modified instead.

Maybe a better name for this property would be "exists", not "deleted"
(maybe the file never existed, so it was never deleted although GtefFile
would say that it has been deleted).