EncodingConverter: call iconv() a last time with NULL inbuf
I've tried to find data for a unit test, without success. To test the
content ending with a shift sequence that gets printed in the outbuf by
the last call of iconv() with the NULL inbuf.

I've found this shift sequence in UTF-7 (from a test in GNU libiconv, in
tests/test-shiftseq.c):
"+2D/"
But nothing is printed in the outbuf.