build: better order of toplevel SUBDIRS
- Regroup source code.
- Have docs/ at the end because gtk-doc is slow. When running `make`, I
can look at the output when it's building the source code, to see if
there is any warning. But for the docs I prefer to do something else in
the meantime, then come back later to see if there is any warning for
the docs (so with the docs at the end, I don't need to scroll up).

(Note, currently building the Gtef docs doesn't take too much time, but
for GtkSourceView for example it takes too much time, so it'll probably
be a problem for Gtef too in the future).