docs: fix FIXME in ActionInfo
It's not really clear what precise vocabulary to use for "save" vs
"win.save". If the target is included, it's the "detailed action name".
But without the target it seems to be just "action name", regardless of
whether it has the GActionMap prefix.

So, provide an example, it's simpler and clearer.

gCSVedit currently doesn't need detailed action names. So keep things
simple, in Gtef the API can anyway be modified later if needed.