Bump API version to 3
The project has been renamed from Gtef to Tepl. But it's still the same
git repo with the same history and the same tags. Setting the API
version to 1 would be confusing, because there are already git tags for
the 1.x.y and 2.x.y versions. Also, it's far less confusing to have the
major package version being equal to the API version.

The package version is set to 2.99.0 because 3.0 is not yet released:
2.99.x are the alpha/beta/rc versions, and 3.0.x will be the stable
version released in September.