amtk: build: drop vapigen
vapigen is confused with Amtk. If someone knows how to fix it, I would
be happy to re-include vapigen.

I've tried the VAPIGEN_GIRDIRS and VAPIGEN_VAPIDIRS options, without
success.

Also I don't know why, but the generated amtk vapi was empty.

Normally the gir is sufficient to use the libraries in Vala, AFAIK.