ApplicationWindow: do not store properties GParamSpec's
With g_object_class_override_property(), we don't have the GParamSpec,
so we cannot call g_object_notify_by_pspec(). For code consistency, it's
better to notify all the properties in the same way, because otherwise
it's dangerous to have some properties that can be notified_by_pspec and
others not (in the same enum).