ApplicationWindow: merge tepl-application-window-actions.c
To update the GAction:enabled properties, listening to some signals is
required, and storing TeplSignalGroup's to the Private struct.

So move the code of tepl-application-window-actions.c to
tepl-application-window.c. With AmtkApplicationWindow, a lot of code has
also been removed from TeplApplicationWindow.

When a lot of GActions will be implemented, maybe TeplApplicationWindow
will become messy and another solution will need to be found.