File saving: write internal _tepl_tab_saving functions
In gedit that code is implemented in GeditTab, but GeditTab is a quite
big class. It's better to delegate the work to other files.

At first I thought about creating a TeplTabSaver class, but actually all
of what is needed is just the async/finish functions, since it's
possible to create a struct for the GTask data.

Let's see how it goes, maybe I'll change my mind later and do
refactorings if needed.

The implementation is currently very simplistic. It'll be improved
later, to handle certain errors by showing different infobars.