build: tests: link Amtk after Tepl
This will be needed for a later commit for the close confirmation
dialog. There was a linker error about undefined amtk symbols when
linking some of the gui tests and unit tests (but not all). This was
caused by a cyclic dependency between several *.o files inside Tepl
(tepl-close-confirm-dialog-single.c calls tepl_application_window
functions, and tepl-application-window.c indirectly depends on
tepl-close-confirm-dialog-single.c, via tepl-tab.c; this is a bit of a
violation of layer, but it is just to add the dialog to the
GtkWindowGroup).

It is anyway a good practice to link the archives in the good order
(I've learned something new about Unix linkers!).