EncodingConverter: fix fixme about errno
See for example the doc of g_strerror(), it recommends to save the value
of errno as soon as possible.

And in the errno(3) manpage, we can read:
"The value of errno is never set to zero  by  any  system
call or library function."
so it means that to recover from an error, we need to set errno to 0
just after saving its value, so that further functions that can
potentially set errno will not return a garbage errno value.