MetadataStore: create class
GLib 2.56 is required for g_file_load_bytes_async()/finish().

An improved version of TeplMetadataManager:
- Async I/O for loading.
- Use the GLib GMarkup parser (not the libxml2).
- Better documentation.

Will write more unit tests later. Some parts may change, like making
get/set_metadata_for_location() private (use it in TeplFileMetadata
only), and I will see when using the metadata API from a real-world
application, to see if the APIs are convenient to use, and if not,
improve them!

It's planned to change the TeplFileMetadata API (which will require to
bump the Tepl API/major version). And of course remove
TeplMetadataManager.

For the unit tests, it would be nice to have a set of XML files. And for
managing that, it's easier with Meson... ;-) (so it's also planned to
migrate Tepl to Meson, finally).