FileMetadata: start to rework the API
This is an API break, so the Tepl API/major version will be bumped in a
forthcoming commit.

GLib 2.64 is required for g_warning_once().

Make TeplFileMetadata independent of TeplFile, to make it more
re-usable. The goal is to expose TeplFileMetadata in the gedit plugin
API.

As a first step, do not care about the fallback to TeplMetadataStore.
Care about GVfs only, and test it well. The integration with
TeplMetadataStore will be done in a later commit.

Remove tepl_file_get_file_metadata(), I prefer not to mix TeplFile with
TeplFileMetadata. The TeplFile* classes are already complex enough. File
metadata will be seen as a separate entity (for now at least, I have the
intention to review the TeplFile API later). TeplFileMetadata will maybe
be associated with TeplBuffer instead (maybe). But there is a desire to
integrate TeplFileMetadata into the framework, somehow; I will do that
later.

Disable the TeplFileMetadata unit tests for now, they will be ported to
the new API in a later commit.