FileMetadata: start to document application use-cases
I went maybe a bit overboard, the text is a bit long, but to have a
quality implementation in apps, thinking about all of this is required.
Hopefully the Tepl framework (the higher-level API, not
TeplFileMetadata) will take care of all of this automatically.
TeplFileMetadata is the toolkit, and as a tool it can be used in many
different ways.