FileMetadata: remove a todo comment
Done for https:// in the unit tests. The unit tests can be extended if
needed.

I will anyway see later how to integrate TeplFileMetadata with the file
loading and saving code of various text editors, and at that point there
will maybe be the need to call the TeplFile mount operation factory
method when loading *metadata* (instead of, or in addition to, the
moment where the *content* is loaded).