View: add new_with_buffer()
It's possible to use TeplView without using TeplBuffer, just a
GtkSourceBuffer. It will ease porting apps that don't use TeplBuffer
yet.

tepl_view_new_with_buffer() was in any case missing, it will be used in
gCSVedit (at least as a transition to use more Tepl features).

TeplView doesn't depend on TeplBuffer. If in the future TeplView depends
on TeplBuffer, then the API will need to be broken.