meson: Check for Visual Studio-specific warning items
We want to force-include msvc_recommended_pragmas.h that is shipped with
Windows builds of GLib so that we can filter out warnings that we don't
really need to care much about, but also let the ones that we should be
wary of stand (error) out.

Also use the -utf-8 compiler flag if it is available, as it avoids
unicode handling issues when attempting to build on East Asian locales.