meson: Fix linking tepl library on Visual Studio
Some of the linker flags cannot be used for Visual Studio, and we can't
just link a DLL entirely from a static library on Visual Studio, as the
linker will optimize things out as it does not see the symbols being
used.

As a result, instead of linking using just the static library, we call
extract_all_objects() on the static library in order to link the final
tepl .so/.dll on all builds.