build: Do not try to enable /utf-8 for MSVC
It turns out that while building on East Asian locales with this flag enabled
creates binaries that do not work as well as binaries that are built with or
without this flag when the locale is switched via Control Panel's "Locale for
non-Unicode Programs" setting to an English locale.

Since the /utf-8 was intended as a convenience to people building on East Asian
locales, there is no point keeping it anymore, to avoid any misconception.