Add TeplLanguageChooser interface
Like GtkSourceStyleSchemeChooser, but for GtkSourceLanguage, to have a
consistent name.

The TeplHighlightModeSelector class will be renamed to
TeplLanguageChooserWidget. And there will also be a Dialog
implementation like in gedit, and maybe also a Button (a Button would be
useful for setting the default language in the preferences dialog, for
example, although the Button implementation doesn't exist in gedit yet).